refactor from flux to flux_calib and move some old routines in flux into deprecated
