+++ pypeit/fluxspec.py
@@ -12,7 +12,7 @@ from astropy import units
 from astropy.io import fits
 
 from pypeit import msgs
-from pypeit.core import flux
+from pypeit.core import flux_calib
 from pypeit.core import load
 from pypeit.core import save
 from pypeit import specobjs
@@ -70,7 +70,7 @@ class FluxSpec(object):
         self.par = par
 
         # Get the extinction data
-        self.extinction_data = flux.load_extinction_data(
+        self.extinction_data = flux_calib.load_extinction_data(
                 self.spectrograph.telescope['longitude'], self.spectrograph.telescope['latitude'])
 
         # Parameters
@@ -156,7 +156,7 @@ class FluxSpec(object):
                 stds = [sobj for sobj in self.std_specobjs if sobj.det == det]
                 if len(stds) == 0:
                     debugger.set_trace()
-                idx = flux.find_standard(stds)
+                idx = flux_calib.find_standard(stds)
                 sv_stds.append(stds[idx])
                 msgs.info("Using standard {} for det={}".format(stds[idx], det))
 
@@ -176,7 +176,7 @@ class FluxSpec(object):
             self.std = std_splice
         elif self.spectrograph.pypeline == 'Echelle':
             # Find brightest object in each order
-            std_brightest = self.std_specobjs[flux.find_standard(self.std_specobjs)]
+            std_brightest = self.std_specobjs[flux_calib.find_standard(self.std_specobjs)]
             std_objid = std_brightest['idx'].split('-')[0]
             self.std_idx = np.zeros(len(self.std_specobjs), dtype=bool)
             for ii in range(len(self.std_specobjs)):
@@ -191,7 +191,7 @@ class FluxSpec(object):
         else:
             # Find brightest object in the exposures
             # Searches over all slits (over all detectors), and all objects
-            self.std_idx = flux.find_standard(self.std_specobjs)
+            self.std_idx = flux_calib.find_standard(self.std_specobjs)
             # Set internal
             self.std = self.std_specobjs[self.std_idx]
             # Step
@@ -365,23 +365,23 @@ class MultiSlit(FluxSpec):
 
         self.sens_dict = {}
         this_wave = self.std.optimal['WAVE']
-        sens_dict_long = flux.generate_sensfunc(this_wave,
-                                               self.std.optimal['COUNTS'],
-                                               self.std.optimal['COUNTS_IVAR'],
-                                               self.std_header['AIRMASS'],
-                                               self.std_header['EXPTIME'],
-                                               self.spectrograph.telescope['longitude'],
-                                               self.spectrograph.telescope['latitude'],
-                                               BALM_MASK_WID=self.par['balm_mask_wid'],
-                                               star_type=self.star_type,
-                                               star_mag=self.star_mag,
-                                               telluric=self.telluric,
-                                               ra=self.std_ra,
-                                               dec=self.std_dec,
-                                               std_file = self.std_file,
-                                               poly_norder=self.poly_norder,
-                                               polycorrect=self.polycorrect,
-                                               debug=self.debug)
+        sens_dict_long = flux_calib.generate_sensfunc(this_wave,
+                                                      self.std.optimal['COUNTS'],
+                                                      self.std.optimal['COUNTS_IVAR'],
+                                                      self.std_header['AIRMASS'],
+                                                      self.std_header['EXPTIME'],
+                                                      self.spectrograph.telescope['longitude'],
+                                                      self.spectrograph.telescope['latitude'],
+                                                      BALM_MASK_WID=self.par['balm_mask_wid'],
+                                                      star_type=self.star_type,
+                                                      star_mag=self.star_mag,
+                                                      telluric=self.telluric,
+                                                      ra=self.std_ra,
+                                                      dec=self.std_dec,
+                                                      std_file = self.std_file,
+                                                      poly_norder=self.poly_norder,
+                                                      polycorrect=self.polycorrect,
+                                                      debug=self.debug)
         self.sens_dict['0'] = sens_dict_long
         self.sens_dict['nslits'] = 1
 
@@ -404,11 +404,11 @@ class MultiSlit(FluxSpec):
         self.load_objs(sci_file, std=False)
         # Run
         for sci_obj in self.sci_specobjs:
-            flux.apply_sensfunc(sci_obj, self.sens_dict['0'], self.sci_header['AIRMASS'],
-                                self.sci_header['EXPTIME'], telluric_correct=self.par['telluric_correct'],
-                                extinct_correct=self.par['extinct_correct'],
-                                longitude=self.spectrograph.telescope['longitude'],
-                                latitude=self.spectrograph.telescope['latitude'])
+            flux_calib.apply_sensfunc(sci_obj, self.sens_dict['0'], self.sci_header['AIRMASS'],
+                                      self.sci_header['EXPTIME'], telluric_correct=self.par['telluric_correct'],
+                                      extinct_correct=self.par['extinct_correct'],
+                                      longitude=self.spectrograph.telescope['longitude'],
+                                      latitude=self.spectrograph.telescope['latitude'])
         self.steps.append(inspect.stack()[0][3])
 
 
@@ -421,7 +421,7 @@ class Echelle(FluxSpec):
         super(Echelle, self).__init__(spectrograph, par, **kwargs)
 
         # Echelle key
-        # TODO add these to the parameters to the parset or try to get rid of these parameters in flux.py
+        # TODO add these to the parameters to the parset or try to get rid of these parameters in flux_calib.py
         self.resolution = 3000. #par['resolution']
         self.nresln = 10.0 #par['nresln']
 
@@ -466,18 +466,18 @@ class Echelle(FluxSpec):
                 this_wave = std.optimal['WAVE_GRID'][wavemask]
 
             counts, ivar = std.optimal['COUNTS'][wavemask], std.optimal['COUNTS_IVAR'][wavemask]
-            sens_dict_iord = flux.generate_sensfunc(this_wave, counts, ivar,
-                                                    float(self.std_header['AIRMASS']),
-                                                    self.std_header['EXPTIME'],
-                                                    self.spectrograph.telescope['longitude'],
-                                                    self.spectrograph.telescope['latitude'],
-                                                    star_type=self.star_type,
-                                                    star_mag=self.star_mag,
-                                                    telluric=self.telluric, ra=self.std_ra, dec=self.std_dec,
-                                                    resolution=self.resolution,
-                                                    BALM_MASK_WID=self.BALM_MASK_WID, std_file=self.std_file,
-                                                    poly_norder=self.poly_norder,
-                                                    polycorrect=self.polycorrect, debug=self.debug)
+            sens_dict_iord = flux_calib.generate_sensfunc(this_wave, counts, ivar,
+                                                          float(self.std_header['AIRMASS']),
+                                                          self.std_header['EXPTIME'],
+                                                          self.spectrograph.telescope['longitude'],
+                                                          self.spectrograph.telescope['latitude'],
+                                                          star_type=self.star_type,
+                                                          star_mag=self.star_mag,
+                                                          telluric=self.telluric, ra=self.std_ra, dec=self.std_dec,
+                                                          resolution=self.resolution,
+                                                          BALM_MASK_WID=self.BALM_MASK_WID, std_file=self.std_file,
+                                                          poly_norder=self.poly_norder,
+                                                          polycorrect=self.polycorrect, debug=self.debug)
             sens_dict_iord['ech_orderindx'] = iord
             self.sens_dict[str(iord)] = sens_dict_iord
         ## add some keys to be saved into primary header in masterframe
@@ -514,10 +514,10 @@ class Echelle(FluxSpec):
             sens_dict_iord = self.sens_dict[str(iord)]
             for sci_obj in self.sci_specobjs:
                 if sci_obj.ech_orderindx == iord:
-                    flux.apply_sensfunc(sci_obj, sens_dict_iord, float(self.sci_header['AIRMASS']),
-                                        self.sci_header['EXPTIME'], extinct_correct=self.par['extinct_correct'],
-                                        longitude=self.spectrograph.telescope['longitude'],
-                                        latitude=self.spectrograph.telescope['latitude'])
+                    flux_calib.apply_sensfunc(sci_obj, sens_dict_iord, float(self.sci_header['AIRMASS']),
+                                              self.sci_header['EXPTIME'], extinct_correct=self.par['extinct_correct'],
+                                              longitude=self.spectrograph.telescope['longitude'],
+                                              latitude=self.spectrograph.telescope['latitude'])
 
         self.steps.append(inspect.stack()[0][3])
