+++ pypeit/metadata.py
@@ -16,7 +16,7 @@ from astropy import table, coordinates, time
 from pypeit import msgs
 from pypeit import utils
 from pypeit.core import framematch
-from pypeit.core import flux
+from pypeit.core import flux_calib
 from pypeit.core import parse
 from pypeit.par import PypeItPar
 from pypeit.par.util import make_pypeit_file
@@ -1310,7 +1310,7 @@ class PypeItMetaData:
     
                 # If an object exists within 20 arcmins of a listed standard,
                 # then it is probably a standard star
-                foundstd = flux.find_standard_file(ra, dec, check=True)
+                foundstd = flux_calib.find_standard_file(ra, dec, check=True)
                 b = self.type_bitmask.turn_off(b, flag='science' if foundstd else 'standard')
     
         # Find the files without any types
